Add multiple iterations test for TokenRateLimitPolicy
